Pdf_web_data_extraction_task/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
idna==3.6
lxml==5.1.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pip==23.2.1
pydantic==2.5.3
pydantic-core==2.14.6
PyMySQL==1.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
tabula-py==2.9.0
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
webencodings==0.5.1
